// src/app/views/pages/apps/components/commons/models/EBankModel.ts
export class EBankModel {
	customerId?: string;
	nationalIssueDate?: string;
	nationalIssuePlace?: string;
	handyNumber?: string;
	email?: string;
	sector?: string;
	address?: string;
	startDate?: string;
	endDate?: string;
	serviceId?: string;
	authenticationType?: string;
	servicePackage?: string;
	serviceStatus?: string;
	userFunction?: string;
	transLimit?: string;
	mainCurrAcc?: string;
	promotionId?: string;
	campaignId?: string;
	salesType?: string;
	salesId?: string;
	brokerType?: string;
	transSeanetBase?: string;
	transSeanetPro?: string;
	coCode?: string;
}